Move files under java folder
