+++ src/java/org/oidc/service/base/HttpArguments.java
@@ -10,7 +10,7 @@ public class HttpArguments {
     /**
      * Specifies whether it is a POST or GET request
      */
-    private HttpMethod method;
+    private HttpMethod httpMethod;
     /**
      * The url of the resource
      */
@@ -26,41 +26,45 @@ public class HttpArguments {
     private HttpHeader header;
 
     /**
-     * @param method Specifies whether it is a POST or GET request
+     * @param httpMethod Specifies whether it is a POST or GET request
      * @param url The url of the resource
      * @param body Used to carry the entity-body associated with the request or response
      *             (optional)
      * @param header Defines the operating parameters of the Http transaction
      */
-    public HttpArguments(HttpMethod method, String url, String body, HttpHeader header) {
-        this.method = method;
+    public HttpArguments(HttpMethod httpMethod, String url, String body, HttpHeader header) {
+        this.httpMethod = httpMethod;
         this.url = url;
         this.body = body;
         this.header = header;
     }
 
     /**
-     * @param method Specifies whether it is a POST or GET request
+     * @param httpMethod Specifies whether it is a POST or GET request
      * @param url The url of the resource
      */
-    public HttpArguments(HttpMethod method, String url) {
-        this.method = method;
+    public HttpArguments(HttpMethod httpMethod, String url) {
+        this.httpMethod = httpMethod;
         this.url = url;
     }
 
     /**
-     * @param method Specifies whether it is a POST or GET request
+     * @param httpMethod Specifies whether it is a POST or GET request
      */
-    public HttpArguments(HttpMethod method) {
-        this.method = method;
+    public HttpArguments(HttpMethod httpMethod) {
+        this.httpMethod = httpMethod;
     }
 
-    public HttpMethod getMethod() {
-        return method;
+    public HttpArguments() {
+
     }
 
-    public void setMethod(HttpMethod method) {
-        this.method = method;
+    public HttpMethod getHttpMethod() {
+        return httpMethod;
+    }
+
+    public void setHttpMethod(HttpMethod httpMethod) {
+        this.httpMethod = httpMethod;
     }
 
     public String getUrl() {
