--- src/org/oidc/common/ClientAuthenticationMethod.java
@@ -1,10 +1,0 @@
-package org.oidc.common;
-
-/**
- * Types of ClientAuthenticationMethods
- */
-public enum ClientAuthenticationMethod {
-    CLIENT_SECRET_BASIC, CLIENT_SECRET_POST,
-    BEARER_HEADER, BEARER_BODY,
-    CLIENT_SECRET_JWT, PRIVATE_KEY_JWT, NONE;
-}
