--- src/org/oidc/common/EndpointName.java
@@ -1,8 +1,0 @@
-package org.oidc.common;
-
-/**
- * The name of the endpoint on the server that the request should be sent to
- */
-public enum EndpointName {
-    AUTHORIZATION, TOKEN, REGISTRATION, USER_INFO;
-}
