--- src/org/oidc/common/FileOrUrl.java
@@ -1,8 +1,0 @@
-package org.oidc.common;
-
-/**
- * Specifies whether KeySpecifications is referring to a file or an url
- */
-public enum FileOrUrl {
-    FILE, URL;
-}
