--- src/org/oidc/common/HttpMethod.java
@@ -1,8 +1,0 @@
-package org.oidc.common;
-
-/**
- * The Http method used to send the request
- */
-public enum HttpMethod {
-    POST, GET;
-}
