--- src/org/oidc/common/JwsAuthenticationMethod.java
@@ -1,8 +1,0 @@
-package org.oidc.common;
-
-/**
- * Types of Jws Authentication methods
- */
-public enum JwsAuthenticationMethod {
-    PRIVATE_KEY_JWT, CLIENT_SECRET_JWT;
-}
