--- src/org/oidc/common/KeyType.java
@@ -1,8 +1,0 @@
-package org.oidc.common;
-
-/**
- * Specifies whether the key is RSA or EC
- */
-public enum KeyType {
-    RSA, EC;
-}
