--- src/org/oidc/common/MissingRequiredAttribute.java
@@ -1,7 +1,0 @@
-package org.oidc.common;
-
-public class MissingRequiredAttribute extends Exception {
-    public MissingRequiredAttribute(String message) {
-        super(message);
-    }
-}
