--- src/org/oidc/common/SerializationType.java
@@ -1,8 +1,0 @@
-package org.oidc.common;
-
-/**
- * The way the request is serialized
- */
-public enum SerializationType {
-    JSON, JWT, URL_ENCODED;
-}
