--- src/org/oidc/common/ServiceName.java
@@ -1,9 +1,0 @@
-package org.oidc.common;
-
-/**
- * The names of different services supported within the Oidcservice layer
- */
-public enum ServiceName {
-    AUTHORIZATION, ACCESS_TOKEN, REFRESH_ACCESS_TOKEN,
-    PROVIDER_INFO_DISCOVERY, WEB_FINGER, REGISTRATION, USER_INFO;
-}
