--- src/org/oidc/common/WebFingerError.java
@@ -1,7 +1,0 @@
-package org.oidc.common;
-
-public class WebFingerError extends Exception {
-    public WebFingerError(String message) {
-        super(message);
-    }
-}
