--- src/org/oidc/service/util/Constants.java
@@ -1,6 +1,0 @@
-package org.oidc.service.util;
-
-public class Constants {
-    public static final String OIDC_ISSUER = "http://openid.net/specs/connect/1.0/issuer";
-    public static final String WEB_FINGER_URL = "https://%s/.well-known/webfinger";
-}
