+++ src/test/java/org/oidc/services/WebfingerTest.java
@@ -12,7 +12,6 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.oidc.common.AddedClaims;
 import org.oidc.common.WebFingerException;
-import org.oidc.service.AbstractService;
 import org.oidc.service.util.Constants;
 
 public class WebfingerTest {
